+++ winter/src/test/kotlin/io/jentz/winter/evaluator/CyclicDependenciesCheckingDirectServiceEvaluatorTest.kt
@@ -1,5 +1,0 @@
-package io.jentz.winter.evaluator
-
-class CyclicDependenciesCheckingDirectServiceEvaluatorTest : AbstractCyclicServiceEvaluatorTest() {
-    override val evaluator = CyclicDependenciesCheckingDirectServiceEvaluator()
-}